backend/requirements.txt
# Django and DRF
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.1

# Database
psycopg2-binary>=2.9.9

# Image processing
Pillow>=10.0.0

# File handling
python-magic>=0.4.27

# ML model requirements
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scikit-learn>=1.3.2

# Medical image processing
pydicom>=2.3.0
nibabel>=4.0.0
opencv-python>=4.8.1.78
pylibjpeg>=1.4.0

# API Testing
pytest>=7.4.3
pytest-django>=4.7.0
coverage>=7.3.2

# Security
python-jose>=3.3.0
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0
gunicorn>=21.2.0
celery>=5.3.6
redis>=5.0.1
flower>=2.0.1